requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
idna==2.8
Jinja2==2.10
livereload==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.0
mkdocs==1.0.4
mkdocs-material==3.1.0
oauthlib==2.1.0
pkg-resources==0.0.0
Pygments==2.3.0
pymdown-extensions==6.0
PySocks==1.6.8
python-dotenv==0.10.1
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
tornado==5.1.1
tweepy==3.7.0
urllib3==1.24.1
